+++ lib/src/widgets/mp_available_scraps_widget.dart
@@ -173,12 +173,12 @@ class _MPAvailableScrapsWidgetState extends State<MPAvailableScrapsWidget> {
 
     Rect? thisBoundingBox = MPInteractionAux.getWidgetRectFromContext(
       widgetContext: context,
-      ancestorGlobalKey: th2FileEditController.thFileWidgetKey,
+      ancestorGlobalKey: th2FileEditController.getTHFileWidgetGlobalKey(),
     );
 
     Rect? childBoundingBox = MPInteractionAux.getWidgetRectFromContext(
       widgetContext: childContext,
-      ancestorGlobalKey: th2FileEditController.thFileWidgetKey,
+      ancestorGlobalKey: th2FileEditController.getTHFileWidgetGlobalKey(),
     );
 
     /// Use the left of this widget and the vertical center of the child (taped