+++ lib/src/widgets/mp_options_edit_overlay_window_widget.dart
@@ -372,12 +372,12 @@ class _MPOptionsEditOverlayWindowWidgetState
   void onOptionSelected(BuildContext childContext, THCommandOptionType type) {
     Rect? thisBoundingBox = MPInteractionAux.getWidgetRectFromContext(
       widgetContext: context,
-      ancestorGlobalKey: th2FileEditController.thFileWidgetKey,
+      ancestorGlobalKey: th2FileEditController.getTHFileWidgetGlobalKey(),
     );
 
     Rect? childBoundingBox = MPInteractionAux.getWidgetRectFromContext(
       widgetContext: childContext,
-      ancestorGlobalKey: th2FileEditController.thFileWidgetKey,
+      ancestorGlobalKey: th2FileEditController.getTHFileWidgetGlobalKey(),
     );
 
     /// Use the right of this widget and the vertical center of the child (taped