+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Infrastructure maintenance:
   * flutter upgrade to 3.38.9.
   * flutter pub upgrade --major-versions
+  * Migrated thFileWidgetKey to getTHFileWidgetGlobalKey() at TH2FileEditController.
 
 ## 0.2.33 - 2026-01-30 - The [_o ano um_](https://renataweber.art/trabalhos/o-ano-um/) release
 * New features: